Update static assert for CHERI architectures.
The Metadata sum field is an Atomic64 this type is defined as a
unitptr_t. Changing Atomic64 to uint64_t (8-bytes) matching the original
static assert has bjsome singificant implications as Atomic64 is assumed
to be the same size of Atomic64 and Atmoic64Location is also assumed to
be the same size as Atomic64. Therefore at the, moment I have retain the
original type definitions and changed the assert.