Preserve capability provenance in addition.
The compiler heuristic takes the left hand side as the provenance source
in the addition. As the code is written this is incorrect. Rather than
reworking the addition I have marked the LHS arguement with the
cheri_no_provenenace argument. This corrects the provenance source and
surpresses the ambiguous provenance source warnings.