Added new option to disable RAW_PTR_EXCLUSION.
The raw_ptr_exclusion attribute is used by a Chromium clang plugin to
enforce rules around usage of raw pointers. On Morello this appears to
cause the following issue:

error: Undefined temporary symbol .L.str.5

This option allows the RAW_PTR_EXCLUSION macro to be disabled, this
appears to solve the issue with temporary symbols.