[cheri] Ensure arg_valuescan store pointer values.
Change TraceArguments arg_values to a a uintptr_t* to ensure that the
arg_values are guaranteed to be wide enough to store a pointer value. On
CHERI architectures this ensures that capability provenance is
preserved, and removes warnings about the cast from a provenance
carrying type to a non-provenanve type.

Fix the trace_arguements_unittest, this was previously causing a failure
in the runtime linker due to the unaligned pointer in the kValues array.