[cheri] Increase metadata to 64B per partition page
With the introduction of capabilities, the PartionSuperPageExtentEntry
no longer fits within the 32B of the metadata per partition page slot.
Therefore, the kPageMetadataShift has been incread to 6 (64B). This,
results in the metadata no longer fitting within a single systemn page
(4KiB).

To account for this the PartitionPage size has been doubled; this
restores the invariants. In various places the code assumes that the
PartitionPage size is 4 times the system page size, these assumptions
have been updated for the larger size.