[cheri] Dummy encode the freelist entry address.
On CHERI architectures encoding the address by reversing the bytes
results in an invalid capability. As the encoding is designed to protect
the address on the freelist from access through buffer overflow attacks,
it is of quesrtionable benefit on CHERI architectures.

Latter versions of partition alloc provide an alternative encoding based
on offsets. Although this looks like it could be adapted for CHERI
architectures, as the value of the approach is currently questionable
this is left for a future stage where the security of partition alloc is
investigated.