[cheri] Temporariliy skip test.
In the tests a span is constructed over start and end values that are
both a sentry. This produces as a CHERI protection fault. As the sentry
values come from a shared object it is not clear that the capabilities
can be rederived. Therefore the test requires some significant rework,
for the time being its is simply skipped.