[cheriabi] Enforce bounds in ParitionAlloc
Initial iteration of enforcing bounds in PartitionAlloc. Bounds are
enforced for allocations, and on freeing, re-allocation the base pool
address is used to widen the bounds of the allocation.

At various points bounds need to be enforce for internal allocator
abstractions, for example, when converting between a ParitionPage and
the metadata entry.

Various tests rely on the ability to access beyond the beyond the bounds
on the allocation. Pragmatic choices have been made as to whether to
disable these parts of the tests or to widen the bounds and allow the
test to succeed.