[cheriabi] Replace max_align_t with 16 for test
As the ipcz test `compile_c_test` is written C `IPCZ_ALIGN` is defined
as `__attribute__((aligned(alignment))`. This requires an expression
rather than a type name.

I have attempted to fix this using `__has_feature(c_alignas)` but this
produced some errors that I wasn't able to get to the bottom of even
looking at the preprocessed output.