Remove FreeBSD patch changes than prevent xcompile
Remove the `!defined(OS_BSD)` conditionals in `Process::CreationTime`.
These conditionals prevent cross-compiling Chromium on Linux host
machines (as the symbols are undefined). Its unclear what the ioriginal
intent here was, but it appears that the original patch resulted in a
`pledge(2)` violation and then rather than removing it was put under an
incorrect preprocessor conditional.